Move Matrix Factorization to algorithms folder
